API-1418: Split into separate files
