+++ src/Akeneo/Connectivity/Connection/back/tests/Integration/Webhook/ProduceAPIEventOnSingleProductCreationIntegration.php
@@ -1,0 +1,50 @@
+<?php
+
+namespace Akeneo\Connectivity\Connection\Tests\Integration\Webhook;
+
+use Akeneo\Pim\Enrichment\Component\Product\Builder\ProductBuilderInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Message\ProductCreated;
+use Akeneo\Platform\Component\EventQueue\BulkEvent;
+use Akeneo\Test\Integration\Configuration;
+use Akeneo\Test\Integration\TestCase;
+use Akeneo\Tool\Component\StorageUtils\Saver\SaverInterface;
+
+/**
+ * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class ProduceAPIEventOnSingleProductCreationIntegration extends TestCase
+{
+    private ProductBuilderInterface $productBuilder;
+    private SaverInterface $productSaver;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->productBuilder = $this->get('pim_catalog.builder.product');
+        $this->productSaver = $this->get('pim_catalog.saver.product');
+    }
+
+    public function test_the_single_product_creation_event()
+    {
+        $product = $this->productBuilder->createProduct('t-shirt');
+        $this->productSaver->save($product);
+
+        $transport = self::$container->get('messenger.transport.business_event');
+
+        $envelopes = $transport->get();
+        $this->assertCount(1, $envelopes);
+
+        /** @var BulkEvent */
+        $bulkEvent = $envelopes[0]->getMessage();
+        $this->assertInstanceOf(BulkEvent::class, $bulkEvent);
+        $this->assertCount(1, $bulkEvent->getEvents());
+        $this->assertContainsOnlyInstancesOf(ProductCreated::class, $bulkEvent->getEvents());
+    }
+
+    protected function getConfiguration(): Configuration
+    {
+        return $this->catalog->useMinimalCatalog();
+    }
+}
