+++ src/Akeneo/Connectivity/Connection/back/tests/Integration/Webhook/ProduceAPIEventOnSingleProductDeletionIntegration.php
@@ -1,0 +1,57 @@
+<?php
+
+namespace Akeneo\Connectivity\Connection\Tests\Integration\Webhook;
+
+use Akeneo\Connectivity\Connection\Tests\CatalogBuilder\Enrichment\ProductLoader;
+use Akeneo\Pim\Enrichment\Component\Product\Message\ProductCreated;
+use Akeneo\Pim\Enrichment\Component\Product\Message\ProductRemoved;
+use Akeneo\Platform\Component\EventQueue\BulkEvent;
+use Akeneo\Test\Integration\Configuration;
+use Akeneo\Test\Integration\TestCase;
+use Akeneo\Tool\Component\StorageUtils\Remover\RemoverInterface;
+
+/**
+ * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class ProduceAPIEventOnSingleProductDeletionIntegration extends TestCase
+{
+    private RemoverInterface $productRemover;
+    private ProductLoader $productLoader;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->productRemover = $this->get('pim_catalog.remover.product');
+        $this->productLoader = $this->get('akeneo_connectivity.connection.fixtures.enrichment.product');
+    }
+
+    public function test_the_single_product_remove_event()
+    {
+        $product = $this->productLoader->create('t-shirt', []);
+        $this->productRemover->remove($product);
+
+        $transport = self::$container->get('messenger.transport.business_event');
+
+        $envelopes = $transport->get();
+        $this->assertCount(2, $envelopes);
+
+        /** @var BulkEvent */
+        $bulkEvent = $envelopes[0]->getMessage();
+        $this->assertInstanceOf(BulkEvent::class, $bulkEvent);
+        $this->assertCount(1, $bulkEvent->getEvents());
+        $this->assertContainsOnlyInstancesOf(ProductCreated::class, $bulkEvent->getEvents());
+
+        /** @var BulkEvent */
+        $bulkEvent = $envelopes[1]->getMessage();
+        $this->assertInstanceOf(BulkEvent::class, $bulkEvent);
+        $this->assertCount(1, $bulkEvent->getEvents());
+        $this->assertContainsOnlyInstancesOf(ProductRemoved::class, $bulkEvent->getEvents());
+    }
+
+    protected function getConfiguration(): Configuration
+    {
+        return $this->catalog->useMinimalCatalog();
+    }
+}
